+++ src/MiNET/MiNET/Entities/Painting.cs
@@ -74,12 +74,12 @@ namespace MiNET.Entities
 			{
 			}
 
-			public override void TakeHit(Entity source, int damage = 1, DamageCause cause = DamageCause.Unknown)
+			public override void TakeHit(Entity source, float damage = 1, DamageCause cause = DamageCause.Unknown)
 			{
 				Entity.DespawnEntity();
 			}
 
-			public override void TakeHit(Entity source, Item tool, int damage = 1, DamageCause cause = DamageCause.Unknown)
+			public override void TakeHit(Entity source, Item tool, float damage = 1, DamageCause cause = DamageCause.Unknown)
 			{
 				Entity.DespawnEntity();
 			}
